requirements.txt
amqp==2.3.2
anyjson==0.3.3
astroid==2.0.4
beautifulsoup4==4.6.3
billiard==3.5.0.4
bs4==0.0.1
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.2.1
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
construct==2.5.3
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.16
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-angular==2.0.0
django-appconf==1.0.2
django-celery==3.2.2
django-celery-results==1.0.4
django-ckeditor-filebrowser-filer==0.2.0b12
django-classy-tags==0.8.0
django-cms==3.4.5
django-compressor==2.2
django-crispy-forms==1.7.2
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.16.3
django-filer==1.3.2
django-filter==1.1.0
django-formtools==2.1
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.5.0
django-ipware==1.1.1
django-js-asset==1.1.0
django-mptt==0.8.7
django-nine==0.1.13
django-parler==1.8
django-paypal==0.5.0
django-polymorphic==1.3
django-post-office==3.0.3
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-sass-processor==0.6
django-sekizai==0.10.0
django-select2==6.0.0
-e git+https://github.com/awesto/django-shop@1e527b1d0e89b6b3585c407cae115e234200b183#egg=django_shop
django-treebeard==4.2.0
djangocms-admin-style==1.2.9
djangocms-bootstrap3==0.3.2
djangocms-cascade==0.16
djangocms-text-ckeditor==3.4.0
djangorestframework==3.7.7
drf-haystack==1.6.1
easy-thumbnails==2.5
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
future==0.17.1
gunicorn==19.9.0
hiredis==0.2.0
html5lib==0.9999999
idna==2.6
ipaddress==1.0.22
isort==4.3.4
jsonfield==1.0.3
kombu==4.2.1
lazy-object-proxy==1.3.1
libsass==0.12.3
mccabe==0.6.1
oauthlib==1.0.3
olefile==0.46
paypal==1.2.5
pefile==2018.8.8
Pillow==4.3.0
pycparser==2.19
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
Pyphen==0.9.5
python-dateutil==2.6.0
python-openid==2.2.5
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2015.7
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.5.0
rjsmin==1.0.12
six==1.11.0
tinycss2==0.6.1
typed-ast==1.1.0
Unidecode==0.4.18
urllib3==1.22
vine==1.1.4
virtualenv==13.1.2
WeasyPrint==43
webencodings==0.5.1
whitenoise==4.1.2
wrapt==1.10.11